Fixed incompatibilities introduced by sigcheck changes

+++ src/txmempool.cpp
@@ -50,14 +50,14 @@ size_t CTxMemPoolEntry::GetTxVirtualSize() const {
     return GetVirtualTransactionSize(nTxSize, sigOpCount);
 }
 
-size_t CTxMemPoolEntry::GetVirtualSizeWithDescendants() const {
+uint64_t CTxMemPoolEntry::GetVirtualSizeWithDescendants() const {
     // note this is distinct from the sum of descendants' individual virtual
     // sizes, and may be smaller.
     return GetVirtualTransactionSize(nSizeWithDescendants,
                                      nSigOpCountWithDescendants);
 }
 
-size_t CTxMemPoolEntry::GetVirtualSizeWithAncestors() const {
+uint64_t CTxMemPoolEntry::GetVirtualSizeWithAncestors() const {
     // note this is distinct from the sum of descendants' individual virtual
     // sizes, and may be smaller.
     return GetVirtualTransactionSize(nSizeWithAncestors,
